Refs #24682: Storyblok - split `generateImageSourcesList` function into two
