+++ components/storyblok/Driver.vue
@@ -34,6 +34,7 @@ import { BaseImage, ImageSourceItem } from 'src/modules/budsies';
 import { Blok } from 'src/modules/vsf-storyblok-module/components';
 
 import getUrlFromLink from './get-url-from-link';
+import generateBreakpointsSpecs from './generate-breakpoints-specs';
 import generateImageSourcesList from './generate-image-sources-list';
 import DriverData from './interfaces/driver-data.interface';
 
@@ -68,11 +69,15 @@ export default (Blok as VueConstructor<InstanceType<typeof Blok> & InjectedServi
         return [];
       };
 
-      return generateImageSourcesList(
+      const breakpointsSpecs = generateBreakpointsSpecs(
         this.itemData.image.filename,
         this.componentWidthCalculator,
-        this.supportsWebp,
         this.itemData.mobile_image.filename
+      )
+
+      return generateImageSourcesList(
+        breakpointsSpecs,
+        this.supportsWebp
       )
     }
   },
