+++ components/storyblok/interfaces/breakpoint-spec.interface.ts
@@ -1,0 +1,5 @@
+export default interface BreakpointSpec {
+  breakpoint: number,
+  width: number,
+  src: string
+}
